Add the One Shot library (#1)
* Add the OneShot module ... including the Shift key

This allows one to hit _and release_ the Control key before another letter to
generate a Control-sequence. See the following project for details:

    https://github.com/keyboardio/Kaleidoscope-OneShot

When a modifier key is pressed, it is shown in a bright white color, helping to
indicate that it is active. Pressed a second time, it is red (the default).

Also, added a README just in case someone stumbles upon it and needs to know
that they probably don't want this ... except for looking at it as yet another
keyboard.io example.